chat_bot_ai/requirements.txt
torch
transformers
sentence-transformers
flask
openai
nltk
scikit-learn
numpy
evaluate
mysql-connector-python
python-dotenv